restructure rs class
- rolling-shutter class now has a method for threading itself and it is in a separate file
